T: Moved std-aware error annotator tests to separate class
