+++ src/test/kotlin/org/rust/ide/annotator/RsErrorAnnotatorTest.kt
@@ -5,7 +5,6 @@
 
 package org.rust.ide.annotator
 
-import com.intellij.testFramework.LightProjectDescriptor
 import org.rust.fileTreeFromText
 
 class RsErrorAnnotatorTest : RsAnnotatorTestBase() {
@@ -589,23 +588,6 @@ class RsErrorAnnotatorTest : RsAnnotatorTestBase() {
             <error descr="Unresolved module">mod foo;</error>
             fn foo() {}
         }
-    """)
-
-    override fun getProjectDescriptor(): LightProjectDescriptor = WithStdlibAndDependencyRustProjectDescriptor
-
-    fun testE0428_RespectsCrateAliases() = checkErrors("""
-        extern crate libc as libc_alias;
-        mod libc {}
-
-        // FIXME: ideally we want to highlight these
-        extern crate alloc;
-        mod alloc {}
-    """)
-
-    fun testE0463_UnknownCrate() = checkErrors("""
-        extern crate alloc;
-
-        <error descr="Can't find crate for `litarvan` [E0463]">extern crate litarvan;</error>
     """)
 
     fun testE0428_IgnoresLocalBindings() = checkErrors("""
